Remove short-cut redundancy check
Doesn't seem to have any performance impact. Actually, this is not quite
true, since usage of `time` shows a performance decrease of about 1% due
to the inclusion of the short-cut. This is probably due to the necessity
of keeping track of the set at each vector operation.

The most important thing about this change is that it may remove a
source of error.