Use unbounded columns
Although the bound is correct (entropies are certainly non-negative), it
may still be a subtle source for errors. The FM elimination algorithm
works on an system of inequalities. Making the non-negativity constraint
known by the redundancy checker on the other hand, may lead to some of
the inequalities being considered redundand and removed. In effect, some
combinations may not be enumerated.